Fix bug suppression commentaire
Fix du bug qui permettez à n'importe qui de supprimer un commentaire sur un ticket.

By Guedesite

+++ Build/controleur/support/ticketCommentaireDelete.php
@@ -1,6 +1,5 @@
 <?php
-if(isset($_Joueur_)) {
-    $pseudo = $_Joueur_['pseudo'];
+if(isset($_Joueur_) && ($ticketCommentaires[$tickets['id']][$i]['auteur'] == $_Joueur_['pseudo'] OR $_Joueur_['rang'] == 1 OR $_PGrades_['PermsDefault']['support']['deleteMemberComm'] == true)) {
     $id_ticket = urldecode($_GET['id_ticket']);
     $id_comm = urldecode($_GET['id_comm']);
     $auteur = urldecode($_GET['auteur']);
@@ -18,22 +17,27 @@ if(isset($_Joueur_)) {
     $get_AuteurCommentaire = $req_AuteurCommentaire->fetch(PDO::FETCH_ASSOC);
 
     $AuteurCommentaire = $get_AuteurCommentaire['auteur'];
-    $ExistCommentaire = $req_ExistCommentaire->rowCount();
-    $ExistTicket = $req_ExistTicket->rowCount();
-    if($ExistTicket == "0") {
-    	header('Location: index.php?&page=support&TicketNotExist=true');
-    } else {
-    	if($ExistCommentaire == "0") {
-            header('Location: index.php?&page=support&CommentaireNotExist=true');
-        } else {
-            if(!$AuteurCommentaire == $pseudo OR !$adminMode = true) {
-             header('Location: index.php?&page=support&SuppressionImpossible=true');
-         } else {
-            $commentairesTickets->DeleteCommentaireTicket($id_comm, $id_ticket, $AuteurCommentaire);
-            header('Location: index.php?&page=support&SuppressionCommentaire=true');
-        }
-    }
-}
+	if($AuteurCommentaire == $_Joueur_['pseudo'] OR $_Joueur_['rang'] == 1 OR $_PGrades_['PermsDefault']['support']['deleteMemberComm'] == true)
+	{
+		$ExistCommentaire = $req_ExistCommentaire->rowCount();
+		$ExistTicket = $req_ExistTicket->rowCount();
+		if($ExistTicket == "0") {
+			header('Location: index.php?&page=support&TicketNotExist=true');
+		} else {
+			if($ExistCommentaire == "0") {
+				header('Location: index.php?&page=support&CommentaireNotExist=true');
+			} else {
+				if(!$AuteurCommentaire == $pseudo OR !$adminMode = true) {
+					 header('Location: index.php?&page=support&SuppressionImpossible=true');
+				 } else {
+					$commentairesTickets->DeleteCommentaireTicket($id_comm, $id_ticket, $AuteurCommentaire);
+					header('Location: index.php?&page=support&SuppressionCommentaire=true');
+				 }
+			}
+		}
+	} else {
+		header('Location: index.php?&page=support&SuppressionCommentaire=false');
+	}
 } else {
     header('Location: index.php?&page=support&NotOnline=true');
 }
